Updated for v9.5.0.70

+++ chilkat/ChilkatLib/Package/include/CkCgi.h
@@ -2,7 +2,7 @@
 //
 //////////////////////////////////////////////////////////////////////
 
-// This header is generated for Chilkat v9.5.0
+// This header is generated for Chilkat 9.5.0.70
 
 #ifndef _CkCgi_H
 #define _CkCgi_H
@@ -49,13 +49,13 @@ class CK_VISIBLE_PUBLIC CkCgi  : public CkMultiByteBase
 	// Properties
 	// ----------------------
 
-	unsigned long get_AsyncBytesRead(void);
+	int get_AsyncBytesRead(void);
 
 
 	bool get_AsyncInProgress(void);
 
 
-	unsigned long get_AsyncPostSize(void);
+	int get_AsyncPostSize(void);
 
 
 	bool get_AsyncSuccess(void);
@@ -82,9 +82,9 @@ class CK_VISIBLE_PUBLIC CkCgi  : public CkMultiByteBase
 	void put_ReadChunkSize(int newVal);
 
 
-	unsigned long get_SizeLimitKB(void);
+	int get_SizeLimitKB(void);
 
-	void put_SizeLimitKB(unsigned long newVal);
+	void put_SizeLimitKB(int newVal);
 
 
 	bool get_StreamToUploadDir(void);
@@ -167,7 +167,7 @@ class CK_VISIBLE_PUBLIC CkCgi  : public CkMultiByteBase
 
 
 
-	unsigned long GetUploadSize(int index);
+	int GetUploadSize(int index);
